Add dataset and ML model training notebook
This commit adds the dataset used for training our machine learning model, as well as the Jupyter notebook used to train the model. The dataset contains 97 samples and was collected from kaggle. The notebook includes the code used to preprocess the data, train the model, and evaluate its performance. We used multiple algorithms to train the model. This work is part of our ongoing project to develop a classifier that can classify individuals into Underweight, Normal and Overweight classes.